requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
geoalchemy2
pydantic
pydantic-settings
aiofiles
